+++ crates/client/rpc/src/providers/mempool.rs
@@ -48,7 +48,6 @@ impl AddTransactionProvider for MempoolAddTxProvider {
         &self,
         declare_v0_transaction: BroadcastedDeclareTransactionV0,
     ) -> RpcResult<DeclareTransactionResult> {
-        log::info!("Checkpoint 2: add_declare_v0_transaction");
         Ok(self.mempool.accept_declare_v0_tx(declare_v0_transaction).map_err(StarknetRpcApiError::from)?)
     }
     async fn add_declare_transaction(