+++ crates/primitives/transactions/src/broadcasted_to_blockifier.rs
@@ -40,17 +40,12 @@ pub fn broadcasted_to_blockifier_v0(
     (blockifier::transaction::transaction_execution::Transaction, Option<ConvertedClass>),
     BroadcastedToBlockifierError,
 > {
-    log::debug!("Checkpoint 4: broadcasted to blockifier v0");
-
     let (class_info, class_hash, extra_class_info) = {
-        let compressed_legacy_class: CompressedLegacyContractClass = (*transaction.contract_class).clone().into();
+        let compressed_legacy_class: CompressedLegacyContractClass = (*transaction.contract_class).clone();
         let class_hash = compressed_legacy_class.compute_class_hash().unwrap();
-        log::debug!("Checkpoint 5: Computed legacy class hash: {:?}", class_hash);
-        let compressed_legacy_class: CompressedLegacyContractClass = (*transaction.contract_class).clone().into();
+        let compressed_legacy_class: CompressedLegacyContractClass = (*transaction.contract_class).clone();
         let class_blockifier =
             compressed_legacy_class.to_blockifier_class().map_err(BroadcastedToBlockifierError::CompilationFailed)?;
-
-        log::debug!("Checkpoint 5: class blockfier sorted");
 
         let class_info = LegacyClassInfo { contract_class: Arc::new(compressed_legacy_class) };
 
@@ -61,13 +56,9 @@ pub fn broadcasted_to_blockifier_v0(
         )
     };
 
-    log::debug!("Checkpoint 6: we have the class info, class hash and the extra class info");
-
     let is_query = transaction.is_query;
     let TransactionWithHash { transaction, hash } =
         TransactionWithHash::from_broadcasted_v0(transaction, chain_id, starknet_version, class_hash);
-
-    log::debug!("Checkpoint 7: tx hash is: {:?}", hash);
 
     let deployed_address = match &transaction {
         Transaction::DeployAccount(tx) => Some(tx.calculate_contract_address()),